requirements.txt
plexapi==4.13.*
confuse~=1.7
pytz~=2022.1
PyYAML==6.0.*
